requirements.txt
boto==2.43.0
dj-database-url==0.4.1
Django==1.10.3
django-push-notifications==1.4.1
django-storages==1.5.1
djangorestframework==3.5.3
gunicorn==19.6.0
Pillow==3.4.2
